Outsource UpdateConfig struct into dedicated file
