+++ cmd/update_test.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
+	"rockstaedt/commit-message-check/internal/model"
 	"testing"
 )
 
@@ -20,7 +21,7 @@ func TestUpdate(t *testing.T) {
 			buffer.Reset()
 			ts := httptest.NewServer(getHandlerFor(`{"tag_name":"v1.0.0"}`))
 			defer ts.Close()
-			config := &UpdateConfig{Version: "v1.0.0", TagUrl: ts.URL}
+			config := &model.UpdateConfig{Version: "v1.0.0", TagUrl: ts.URL}
 
 			status := Update(config)
 
@@ -32,7 +33,7 @@ func TestUpdate(t *testing.T) {
 			ts := httptest.NewServer(getHandlerFor(`{"tag_name":"v1.1.0"}`))
 			defer ts.Close()
 			tempDir := t.TempDir()
-			config := &UpdateConfig{Version: "v1.0.0", TagUrl: ts.URL, DownloadPath: tempDir}
+			config := &model.UpdateConfig{Version: "v1.0.0", TagUrl: ts.URL, DownloadPath: tempDir}
 
 			_ = Update(config)
 
@@ -44,7 +45,7 @@ func TestUpdate(t *testing.T) {
 		buffer.Reset()
 		ts := httptest.NewServer(getHandlerFor("", 500))
 		defer ts.Close()
-		config := &UpdateConfig{Version: "v1.0.0", TagUrl: ts.URL, DownloadPath: ""}
+		config := &model.UpdateConfig{Version: "v1.0.0", TagUrl: ts.URL, DownloadPath: ""}
 
 		status := Update(config)
 
@@ -112,7 +113,7 @@ func TestDownloadScript(t *testing.T) {
 			}
 		}))
 		defer ts.Close()
-		config := &UpdateConfig{LatestVersion: "v1.1.1", DownloadPath: tempDir, BinaryBaseUrl: ts.URL}
+		config := &model.UpdateConfig{LatestVersion: "v1.1.1", DownloadPath: tempDir, BinaryBaseUrl: ts.URL}
 
 		status := downloadScript(config)
 
@@ -123,7 +124,7 @@ func TestDownloadScript(t *testing.T) {
 	})
 
 	t.Run("returns 1 when error at creating file", func(t *testing.T) {
-		config := &UpdateConfig{DownloadPath: getProtectedPath(t)}
+		config := &model.UpdateConfig{DownloadPath: getProtectedPath(t)}
 
 		status := downloadScript(config)
 
@@ -132,7 +133,7 @@ func TestDownloadScript(t *testing.T) {
 
 	t.Run("return 2 when http protocol error", func(t *testing.T) {
 		tempDir := t.TempDir()
-		config := &UpdateConfig{DownloadPath: tempDir, BinaryBaseUrl: "/xxx"}
+		config := &model.UpdateConfig{DownloadPath: tempDir, BinaryBaseUrl: "/xxx"}
 
 		status := downloadScript(config)
 
@@ -145,7 +146,7 @@ func TestDownloadScript(t *testing.T) {
 			w.WriteHeader(500)
 		}))
 		defer ts.Close()
-		config := &UpdateConfig{DownloadPath: tempDir, BinaryBaseUrl: ts.URL}
+		config := &model.UpdateConfig{DownloadPath: tempDir, BinaryBaseUrl: ts.URL}
 
 		status := downloadScript(config)
