+++ internal/model/update_config.go
@@ -1,0 +1,9 @@
+package model
+
+type UpdateConfig struct {
+	Version       string
+	LatestVersion string
+	TagUrl        string
+	BinaryBaseUrl string
+	DownloadPath  string
+}
