+++ main.go
@@ -7,6 +7,7 @@ import (
 	"log"
 	"os"
 	"rockstaedt/commit-message-check/cmd"
+	"rockstaedt/commit-message-check/internal/model"
 	"rockstaedt/commit-message-check/util"
 )
 
@@ -52,7 +53,7 @@ func main() {
 	case "uninstall":
 		status = cmd.Uninstall(gitPath)
 	case "update":
-		config := &cmd.UpdateConfig{
+		config := &model.UpdateConfig{
 			Version:       version,
 			TagUrl:        "https://api.github.com/repos/rockstaedt/commit-message-check/releases/latest",
 			BinaryBaseUrl: "https://github.com/rockstaedt/commit-message-check/releases/latest/download/",
